requirements.txt

flask
mysql-connector-python
openai
google-generativeai
